backend/requirements.txt
Django==3.2.16
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
Pillow==10.3.0
psycopg2-binary==2.9.3 
python-dotenv==1.0.1
requests==2.26.0
